+++ python/glide-sync/glide_sync/__init__.py
@@ -69,6 +69,8 @@ from glide_shared import (
     GeoSearchCount,
     GeospatialData,
     GeoUnit,
+    GlideClientConfiguration,
+    GlideClusterClientConfiguration,
     GlideError,
     HashFieldConditionalChange,
     IamAuthConfig,
@@ -153,12 +155,17 @@ from glide_shared import (
     json_batch,
 )
 
-from .config import GlideClientConfiguration, GlideClusterClientConfiguration
 from .glide_client import GlideClient, GlideClusterClient, TGlideClient
 from .logger import Level as LogLevel
 from .logger import Logger
 from .opentelemetry import OpenTelemetry
-from .sync_commands import ft, glide_json
+from .sync_commands import (
+    ALL_CHANNELS,
+    ALL_PATTERNS,
+    ALL_SHARDED_CHANNELS,
+    ft,
+    glide_json,
+)
 from .sync_commands.cluster_scan_cursor import ClusterScanCursor
 from .sync_commands.script import Script
 from .sync_commands.utils import get_min_compressed_size
@@ -275,6 +282,9 @@ __all__ = [
     "UpdateOptions",
     # PubSub
     "PubSubMsg",
+    "ALL_CHANNELS",
+    "ALL_PATTERNS",
+    "ALL_SHARDED_CHANNELS",
     # Json
     "glide_json",
     "json_batch",